+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.21.0] - 2026-09-14
+
 ### Breaking
 
 - **`context.client.data` no longer compiles in a collection callback.** A